requirements.txt
qdrant-client>=1.7.0,<2.0.0
openai>=1.6.0,<2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0,<1.0.0
tqdm>=4.66.0